% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getDayAheadAuctionEEX}
\alias{getDayAheadAuctionEEX}
\title{Two possible sources: from EEX website or from EPEX SPOT Website. EPEX SPOT has few days on one site EEX only one day}
\usage{
getDayAheadAuctionEEX(startDate, endDate, product)
}
\description{
example link for 2017-05-25 for german/austrian market: https://www.epexspot.com/en/market-data/dayaheadauction/auction-table/2017-05-25/DE/24
}
