% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getDayAheadAuctionEPEXSPOT}
\alias{getDayAheadAuctionEPEXSPOT}
\title{getDayAheadAuctionEPEXSPOT}
\usage{
getDayAheadAuctionEPEXSPOT(startDate, endDate, product = "60",
  country = "DE")
}
\arguments{
\item{startDate}{- Set the start date for the price data period}

\item{endDate}{- Set the end date for the price data period}

\item{product}{- Sets which product should be crawled. There is hourly price data ("60"), block price data ("block") and the base and peak load price data ("base", "peak") data. Default value is "60" for the hourly data.}

\item{country}{- Defines the country from which the data should be crawled. Default value is "DE". There is also "FR" (France) and "CH" (Swiss)}
}
\value{
a data.frame with DateTime as POSIXct object and Last prices of hourly data.
}
\description{
This function returns the price data of the EPEX SPOT Day-Ahead-Auction for a time period.
For french, german (Phelix) and swiss (swissix) --> MCC = Market Coulped Contracts??
https://www.epexspot.com/en/market-data/dayaheadauction

Always get 7 days on one website (request). Date in request link is the latest date.
}
\examples{
h <- getDayAheadAuctionEPEXSPOT("2017-05-20", "2017-05-26", "60")

}
