% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getIntradayAuctionEPEXSPOT}
\alias{getIntradayAuctionEPEXSPOT}
\title{getIntradayAuctionEPEXSPOT}
\usage{
getIntradayAuctionEPEXSPOT(startDate, endDate)
}
\arguments{
\item{startDate}{- Set the start date for the price data period}

\item{endDate}{- Set the end date for the price data period}
}
\value{
a data.frame with DateTime as POSIXct object and intraday auction price data of the given product. The columns are DateTime and the 15min Prices with the Volume as well as daily data of: OffPeak, OffPeak1, SunPeak, OffPeak2, BasePrice", BaseVolume, PeakPrice, PeakVolume
}
\description{
This function returns the price data of the EPEX SPOT Intraday Auction for a time period.
Only for German Market. Only 15min data, bock prices with base and peak
https://www.epexspot.com/en/market-data/intradayauction

Always get 7 days on one website (request). Date in request link is the latest date.
}
\examples{
h <- getIntradayAuctionEPEXSPOT("2017-05-20", "2017-05-26")

}
