% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getPHELIXDEFuturesForADate}
\alias{getPHELIXDEFuturesForADate}
\title{Helper function for @seealso getPHELIXDEFutures}
\usage{
getPHELIXDEFuturesForADate(date, product)
}
\description{
Helper function for @seealso getPHELIXDEFutures
}
