% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{parseDAAEPEXSPOT}
\alias{parseDAAEPEXSPOT}
\title{Helper function for @seealso getDayAheadAuctionEPEXSPOT}
\usage{
parseDAAEPEXSPOT(htmlDoc, country, latestDate)
}
\description{
Helper function for @seealso getDayAheadAuctionEPEXSPOT
}
