% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{parseIAEPEXSPOT}
\alias{parseIAEPEXSPOT}
\title{Helper function for @seealso getIntradayAuctionEPEXSPOT}
\usage{
parseIAEPEXSPOT(htmlDoc, latestDate)
}
\description{
Helper function for @seealso getIntradayAuctionEPEXSPOT
}
