% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{qplotIntradayContinuous}
\alias{qplotIntradayContinuous}
\title{qplotIntradayContinuous}
\usage{
qplotIntradayContinuous(df)
}
\arguments{
\item{df}{- The input data.frame with a few days of cont. intr. trading data (Low,High,Last and DateTimes)}
}
\value{
a pre-defined quick plot with Price and volume data (Low, High, Last in €/MWh) and Volume in GWh (Buy and Sell). No plot object!
}
\description{
This qick plot (qplot...) function visualizes the Lows, Highs and Last prices as well as their buy and sell volume of a given data.frame of continuous intraday trading data of the EPEX SPOT @seealso getIntradayContinuousEPEXSPOT
The plot function makes only sense to visualize a few days. Otherwise the visual elements are to small
}
\examples{
# get the input data
df <- getIntradayContinuousEPEXSPOT("2017-05-24", "2017-05-26", "15", "DE")

plot <- qplotContIntraEPEXSPOT(df)
plot

}
